refactor: extract parse_arguments into cli/parser.py
Move the CLI argument parser (~1,250 lines) from generator.py to
its canonical location in cli/parser.py. Also moves four constants
(QUALITY_SEVERITY_ORDER, QUALITY_SEVERITY_RANK, DEFAULT_AUTO_PRUNE_*)
to core/constants.py. generator.py re-exports for backwards compat.

Reduces generator.py from ~15,600 to ~14,350 lines.
